refactor: separate all email providers into individual packages
