+++ providers/gmail/gmail_email_sender_test.go
@@ -1,4 +1,4 @@
-package providers
+package gmail
 
 import (
 	"bytes"
