+++ providers/helper_functions.go
@@ -9,7 +9,7 @@ import (
 	"github.com/darkrockmountain/gomail"
 )
 
-// buildMimeMessage constructs the MIME message for the email, including text, HTML, and attachments.
+// BuildMimeMessage constructs the MIME message for the email, including text, HTML, and attachments.
 // This function builds a multipart MIME message based on the provided email message. It supports plain text,
 // HTML content, and multiple attachments.
 //
@@ -35,12 +35,12 @@ import (
 //	        },
 //	    },
 //	}
-//	mimeMessage, err := buildMimeMessage(message)
+//	mimeMessage, err := BuildMimeMessage(message)
 //	if err != nil {
 //	    log.Fatalf("Failed to build MIME message: %v", err)
 //	}
 //	fmt.Println(string(mimeMessage))
-func buildMimeMessage(message gomail.EmailMessage) ([]byte, error) {
+func BuildMimeMessage(message gomail.EmailMessage) ([]byte, error) {
 	var msg bytes.Buffer
 
 	// Determine boundaries
@@ -127,7 +127,7 @@ func buildMimeMessage(message gomail.EmailMessage) ([]byte, error) {
 	return msg.Bytes(), nil
 }
 
-// strPtr takes a string value and returns a pointer to that string.
+// StrPtr takes a string value and returns a pointer to that string.
 // This function is useful when you need to work with string pointers, such as in
 // scenarios where you need to pass a string by reference or handle optional string fields.
 //
@@ -140,12 +140,12 @@ func buildMimeMessage(message gomail.EmailMessage) ([]byte, error) {
 // Example usage:
 //
 //	name := "John Doe"
-//	namePtr := strPtr(name)
+//	namePtr := StrPtr(name)
 //	fmt.Println(namePtr)  // Output: memory address of the string
 //	fmt.Println(*namePtr) // Output: "John Doe"
 //
 // Detailed explanation:
-// The strPtr function creates a pointer to the given string `str`.
+// The StrPtr function creates a pointer to the given string `str`.
 // This can be particularly useful in the following scenarios:
 //  1. Passing strings by reference to functions, which can help avoid copying large strings.
 //  2. Working with data structures that use pointers to represent optional fields or nullable strings.
@@ -153,6 +153,6 @@ func buildMimeMessage(message gomail.EmailMessage) ([]byte, error) {
 //
 // By using this function, you can easily obtain a pointer to a string and utilize it in contexts
 // where pointers are needed, thus enhancing flexibility and efficiency in your Go programs.
-func strPtr(str string) *string {
+func StrPtr(str string) *string {
 	return &str
 }
