+++ providers/mailgun/mailgun_email_sender_test.go
@@ -1,4 +1,4 @@
-package providers
+package mailgun
 
 import (
 	"context"
