+++ providers/microsoft365/microsoft365_email_sender_test.go
@@ -1,4 +1,4 @@
-package providers
+package microsoft365
 
 import (
 	"bytes"
