+++ providers/sparkpost/sparkpost_email_sender.go
@@ -1,4 +1,4 @@
-package providers
+package sparkpost
 
 import (
 	"fmt"
